Reduced cooldown for commands
Reduced the cooldown for commands as people were trying to do the
command but were unable to due to it recently being used however it
wasn't visible above due to it deleting 120 seconds later. Next update
possibly doing a timed cooldown instead. Command cooldowns are now half
of the message timer cooldown.